requirements.txt
openai>=1.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=2.0.0
Pillow>=9.0.0
videodb>=0.1.0
